sphinx: create new crypto.go file
