+++ sphinx.go
@@ -9,18 +9,12 @@ import (
 	"io"
 	"math/big"
 
-	"github.com/aead/chacha20"
 	"github.com/btcsuite/btcd/btcec"
 	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcutil"
 )
 
 const (
-	// HMACSize is the length of the HMACs used to verify the integrity of
-	// the onion. Any value lower than 32 will truncate the HMAC both
-	// during onion creation as well as during the verification.
-	HMACSize = 32
-
 	// addressSize is the length of the serialized address used to uniquely
 	// identify the next hop to forward the onion to. BOLT 04 defines this
 	// as 8 byte channel_id.
@@ -74,14 +68,6 @@ const (
 	// baseVersion represent the current supported version of onion packet.
 	baseVersion = 0
 )
-
-// Hash256 is a statically sized, 32-byte array, typically containing
-// the output of a SHA256 hash.
-type Hash256 [sha256.Size]byte
-
-// zeroHMAC is the special HMAC value that allows the final node to determine
-// if it is the payment destination or not.
-var zeroHMAC [HMACSize]byte
 
 // OnionPacket is the onion wrapped hop-to-hop routing information necessary to
 // propagate a message through the mix-net without intermediate nodes having
@@ -476,107 +462,6 @@ func (f *OnionPacket) Decode(r io.Reader) error {
 	return nil
 }
 
-// calcMac calculates HMAC-SHA-256 over the message using the passed secret key
-// as input to the HMAC.
-func calcMac(key [keyLen]byte, msg []byte) [HMACSize]byte {
-	hmac := hmac.New(sha256.New, key[:])
-	hmac.Write(msg)
-	h := hmac.Sum(nil)
-
-	var mac [HMACSize]byte
-	copy(mac[:], h[:HMACSize])
-
-	return mac
-}
-
-// xor computes the byte wise XOR of a and b, storing the result in dst.
-func xor(dst, a, b []byte) int {
-	n := len(a)
-	if len(b) < n {
-		n = len(b)
-	}
-	for i := 0; i < n; i++ {
-		dst[i] = a[i] ^ b[i]
-	}
-	return n
-}
-
-// generateKey generates a new key for usage in Sphinx packet
-// construction/processing based off of the denoted keyType. Within Sphinx
-// various keys are used within the same onion packet for padding generation,
-// MAC generation, and encryption/decryption.
-func generateKey(keyType string, sharedKey *Hash256) [keyLen]byte {
-	mac := hmac.New(sha256.New, []byte(keyType))
-	mac.Write(sharedKey[:])
-	h := mac.Sum(nil)
-
-	var key [keyLen]byte
-	copy(key[:], h[:keyLen])
-
-	return key
-}
-
-// generateCipherStream generates a stream of cryptographic psuedo-random bytes
-// intended to be used to encrypt a message using a one-time-pad like
-// construction.
-func generateCipherStream(key [keyLen]byte, numBytes uint) []byte {
-	var (
-		nonce [8]byte
-	)
-	cipher, err := chacha20.NewCipher(nonce[:], key[:])
-	if err != nil {
-		panic(err)
-	}
-	output := make([]byte, numBytes)
-	cipher.XORKeyStream(output, output)
-
-	return output
-}
-
-// computeBlindingFactor for the next hop given the ephemeral pubKey and
-// sharedSecret for this hop. The blinding factor is computed as the
-// sha-256(pubkey || sharedSecret).
-func computeBlindingFactor(hopPubKey *btcec.PublicKey,
-	hopSharedSecret []byte) Hash256 {
-
-	sha := sha256.New()
-	sha.Write(hopPubKey.SerializeCompressed())
-	sha.Write(hopSharedSecret)
-
-	var hash Hash256
-	copy(hash[:], sha.Sum(nil))
-	return hash
-}
-
-// blindGroupElement blinds the group element P by performing scalar
-// multiplication of the group element by blindingFactor: blindingFactor * P.
-func blindGroupElement(hopPubKey *btcec.PublicKey, blindingFactor []byte) *btcec.PublicKey {
-	newX, newY := btcec.S256().ScalarMult(hopPubKey.X, hopPubKey.Y, blindingFactor[:])
-	return &btcec.PublicKey{btcec.S256(), newX, newY}
-}
-
-// blindBaseElement blinds the groups's generator G by performing scalar base
-// multiplication using the blindingFactor: blindingFactor * G.
-func blindBaseElement(blindingFactor []byte) *btcec.PublicKey {
-	newX, newY := btcec.S256().ScalarBaseMult(blindingFactor)
-	return &btcec.PublicKey{btcec.S256(), newX, newY}
-}
-
-// generateSharedSecret generates the shared secret for a particular hop. The
-// shared secret is generated by taking the group element contained in the
-// mix-header, and performing an ECDH operation with the node's long term onion
-// key. We then take the _entire_ point generated by the ECDH operation,
-// serialize that using a compressed format, then feed the raw bytes through a
-// single SHA256 invocation.  The resulting value is the shared secret.
-func generateSharedSecret(pub *btcec.PublicKey, priv *btcec.PrivateKey) Hash256 {
-	s := &btcec.PublicKey{}
-	x, y := btcec.S256().ScalarMult(pub.X, pub.Y, priv.D.Bytes())
-	s.X = x
-	s.Y = y
-
-	return sha256.Sum256(s.SerializeCompressed())
-}
-
 // ProcessCode is an enum-like type which describes to the high-level package
 // user which action should be taken after processing a Sphinx packet.
 type ProcessCode int
@@ -798,16 +683,6 @@ func unwrapPacket(onionPkt *OnionPacket, sharedSecret *Hash256,
 	return innerPkt, &hopData, nil
 }
 
-// sharedSecretGenerator is an interface that abstracts away exactly *how* the
-// shared secret for each hop is generated.
-//
-// TODO(roasbef): rename?
-type sharedSecretGenerator interface {
-	// generateSharedSecret given a public key, generates a shared secret
-	// using private data of the underlying sharedSecretGenerator.
-	generateSharedSecret(dhKey *btcec.PublicKey) (Hash256, error)
-}
-
 // processOnionPacket performs the primary key derivation and handling of onion
 // packets. The processed packets returned from this method should only be used
 // if the packet was not flagged as a replayed packet.
@@ -845,20 +720,6 @@ func processOnionPacket(onionPkt *OnionPacket, sharedSecret *Hash256,
 		ForwardingInstructions: *outerHopData,
 		NextPacket:             innerPkt,
 	}, nil
-}
-
-// generateSharedSecret generates the shared secret by given ephemeral key.
-func (r *Router) generateSharedSecret(dhKey *btcec.PublicKey) (Hash256, error) {
-	var sharedSecret Hash256
-
-	// Ensure that the public key is on our curve.
-	if !btcec.S256().IsOnCurve(dhKey.X, dhKey.Y) {
-		return sharedSecret, ErrInvalidOnionKey
-	}
-
-	// Compute our shared secret.
-	sharedSecret = generateSharedSecret(dhKey, r.onionKey)
-	return sharedSecret, nil
 }
 
 // Tx is a transaction consisting of a number of sphinx packets to be atomically
